[refactor] the code
