+++ App/Course App/Views/UIKit/UICollectionViewCells/LabelCollectionViewCell.swift
@@ -16,6 +16,7 @@ final class LabelCollectionViewCell: UICollectionViewCell, ReusableIdentifier {
         setupUI()
     }
     
+    @available(*, unavailable)
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
